backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
python-dotenv==0.19.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
pytest==7.4.2
black==23.9.1
flake8==6.1.0
openai==1.3.0
httpx==0.24.1
mysqlclient==2.1.1
PyMySQL==1.0.2
sentence-transformers==2.2.2
numpy==1.24.3 